+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/plugin_scripts_js/JavaScriptBridgeJS.java
@@ -220,7 +220,7 @@ public class JavaScriptBridgeJS {
           "      var _callHandlerID = setTimeout(function(){});" +
           "      window." + JAVASCRIPT_BRIDGE_NAME + "._callHandler(arguments[0], _callHandlerID, JSON.stringify(Array.prototype.slice.call(arguments, 1)));" +
           "      return new Promise(function(resolve, reject) {" +
-          "        window." + JAVASCRIPT_BRIDGE_NAME + "[_callHandlerID] = resolve;" +
+          "        window." + JAVASCRIPT_BRIDGE_NAME + "[_callHandlerID] = {resolve: resolve, reject: reject};" +
           "      });" +
           "    };" +
           "  }"+
@@ -230,7 +230,7 @@ public class JavaScriptBridgeJS {
           "    var _callHandlerID = setTimeout(function(){});" +
           "    window.top." + JAVASCRIPT_BRIDGE_NAME + "._callHandler(arguments[0], _callHandlerID, JSON.stringify(Array.prototype.slice.call(arguments, 1)));" +
           "    return new Promise(function(resolve, reject) {" +
-          "      window.top." + JAVASCRIPT_BRIDGE_NAME + "[_callHandlerID] = resolve;" +
+          "      window.top." + JAVASCRIPT_BRIDGE_NAME + "[_callHandlerID] = {resolve: resolve, reject: reject};" +
           "    });" +
           "  };" +
           "}" +