+++ example/integration_test/chrome_safari_browser/open_and_close.dart
@@ -45,7 +45,7 @@ void openAndClose() {
             activityButton: ActivityButton(
                 templateImage: UIImage(systemName: "sun.max"),
                 extensionIdentifier:
-                    "com.pichillilorenzo.flutter-inappwebview6-example.test")));
+                    "com.pichillilorenzo.flutter-inappwebview-example7.test")));
     await chromeSafariBrowser.opened.future;
     expect(chromeSafariBrowser.isOpened(), true);
     expect(() async {