+++ lib/src/in_app_browser/in_app_browser.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 import 'dart:collection';
 import 'dart:typed_data';
+import 'dart:ui';
 
 import 'package:flutter/services.dart';